Added Mocking and Test Driven Development - Lab

+++ 16.MockingAndTestDrivenDevelopment-Lab/INStock/Contracts/IProductStock.cs
@@ -20,9 +20,9 @@
 
         IProduct FindMostExpensiveProduct();
 
-        IEnumerable<IProduct> FindAllInRange(double lo, double hi);
+        IEnumerable<IProduct> FindAllInRange(decimal lo, decimal hi);
 
-        IEnumerable<IProduct> FindAllByPrice(double price);
+        IEnumerable<IProduct> FindAllByPrice(decimal price);
 
         IEnumerable<IProduct> FindAllByQuantity(int quantity);
     }
